requirements.txt
fastapi==0.62.0
uvicorn==0.12.3
tortoise-orm[asyncpg]==0.16.18
